requirements.txt
pandas == 1.3.3
Pillow == 8.3.2
scikit-learn == 0.24.1
streamlit ==  1.9.0
numpy == 1.21.2